PHYSLITE: Adding test for triggers in metadata
Adding a test that the triggers we are interested in are included in the
AOD metadata. If they are, then they should be safe to use for matching.
If they are not, then it could be the case (for example) that we are
running on a file that is MC16a and our trigger is a 2018 trigger.