Use average interactions per crossing from the lumi conditions data.
Use the average interactions per crossing from the luminosity conditions
data for the TRT drift circle mu corrections. This works around the problem
that the average interactions per crossing is currently not set in the xAOD
event info.

The mu correction is current disabled in the configuration to avoid output
changes.