BPHY8: Update of options and settings for isolation and close track tools
This is an update of BPHY8 in preparation of the derivation production
for the full Run 2 B(s)mumu analysis.  It mainly moves improvements in
the calculation of isolation- and close-track-related variables from the
ntuple maker level to the derivation level.  These optimizations
(and extensions) to the existing tools were studied on ntuple level
using a small subset of special DAODs in which all ID track
information was kept.  However, in order to reduce the size of the
BPHY8 DAODs again to a reasonable level before a derivation production
run on all Run 2 data, it is mandatory to apply the necessary
modifications to the corresponding tool classes on derivation level.

The derivation-level code has been validated against the existing,
optimized ntuple-level code using small locally produced DAOD files.
Once this BPHY8 code and settings update will be included in a
new derivation cache, a derivation production request for a small
subset of runs is forseen in order to conclude the validation
with somewhat larger statistics.

Details on the modifications are:

Update of options for B and muon isolation as well as close track tools:
- New track-to-PV association types ("track type options") for the
  track selection requirements;
- Option to use the TrackVertexAssociationTool for track-to-PV
  assignment.
- Addition of simple track-to-vertex chi2 calculation method in order
  to be able to compare it with the same from the ntuple maker level
  (used for validation purposes);
- Inclusion of the vertex uncertainties into the calculation of the
  track-to-vertex chi2 for isolation and close-track-related variables;
- Added code to debug the track-to-vertex assignment and the calculation
  of the chi2 of the track-to-SV distance to the
  BVertexClosestTrackTool (off by default);

Update of options for the AugOriginalCounts tool:
- Add original number of tracks as decoration to a primary vertex.
  (Which is no longer accessible once ID tracks are removed from the DAOD.)
- Add sqrt(sum(pt^2)) as decoration option for primary vertices.
  (Also no longer calculatable using the DAOD once (most) ID tracks
   have been removed from the DAOD.)