Also testing for SimulateCavern for EtaPhiFilter
At Sven's request, adding a test for the simulate cavern flag before
adding the eta/phi filter. I can't think of any situation in which we
would want to simulate the cavern but we would not want to simulate
forwward particles. After discussion with John, he agreed with the
addition.