TrigEgammaAnalysisTools: Fix property initialization ordering problem
TrigEgammaAnalysisBaseTool holds a plotting tool, m_plot.  It also has
a property doEmulation, and it sets up a callback on this property so that
when it it set, it is propagated to the subordinate tools, including
m_plot.  However, if the JO service ends up setting the doEmulation property
before the PlotTool property, then m_plot will not be valid in the callback,
leading to errors from ToolSvc.

To fix this, check that m_plot is valid in the callback, and further
arrange to call the call back in initialize().