Dropped several components of the Metadata handling infrastructure
The following classes have been dropped:
  GenericMetadataTool
  GenericMetadataToolNoAux
  MetaInputLoader
  WriteMetaHandle

These components are currently not being used by our code, except for some tests.
The idea is to temporarily drop them in order to facilitate the migration
of metadata handling code to MetaContainers and ReadMetaHandles. Some of the
dropped components may be reintroduced later, as needed.