PyJobTransforms: More py3 unicode fixes
Athena emits some log messages that contain unicode characters.
With python 3, this will causes crashes in trfExe if the logging
sinks aren't set up with the proper encoding.  Depending on the
setting of LANG, they may not be.

Explicitly open logging files with utf-8 encoding.
If sys.stdout is using an ascii/ansi encoding, then open a new stream
with the same fileno with utf-8 encoding.