Fixed a segfault in FileMetaDataCreatorTool::update() in an ESDtoDPD reco test
The sgfault started to show up after merging !28081. As a result of this MR,
FileMetaDataCreatorTool::beginInputFile() was no longer called, and this lead
to a segfault later on inside FileMetaDataCreatorTool::update().

The segfault was fixed by forwarding FileMetaDataCreatorTool::beginInputFile(SG::SourceID&)
(which is called by MetaDataSvc) to FileMetaDataCreatorTool::beginInputFile().
A similar trick had already been implemented for FileMetaDataTool.

Fixes the crash reported in ATLASRECTS-5257