Restructuring config data classes
Prior all DataStructure-derived classes were holding a copy of the
ptree that they were constructed from. Now they hold either a
reference (ptree*) or own the original (shared_ptr<ptree>).

Basically the objects L1Menu, HLTMenu, and L1-/HLTPrescaleSet will hold
the ptree created from the json files. All sub structures will only have
a pointer to the sub-tree which is part of the big ptree.

This avoids copying ptree's around and inflating the size of the menu.
One has to be carefull when creating the sub-structure and only pass
references to sub-trees of the original tree to the constructors.