AthenaMonitoringKernel: Fixes for unit test
The GenericMonFilling_test used the non-const `FindBin` function to test
for the existence of bins with a certain label. However, this function
actually creates the label if it does not exist. Switch to using the
`FindFixBin` function that does not have this side-effect. Also make the
`getHist` function return a `const` pointer to avoid such problems in
the future.

Fixed also one place where the wrong axis was being required and due to
the above, this was not being caught by the test.