Reordering additions in PHYS and PHYSLITE
In order to ensure that all the information is available when SeqPHYS or
SeqPHYSLITE is run, we need to shift the SeqPHYS* later in the
derivation job. This is in part because quite a few truth tools add
themselves to the 'main' sequence (in order to avoid repetition of
work), and those truth tools need to run before the SeqPHYS* tools that
look for their output.