PyDumper: Fix problem with pixel configuration
If we run the event dump job configuration, it was crashing
because PixelDCSCondHVAlg was being added to the conditions sequence
without the stuff it depends on.  However, we don't need any of that
stuff at all here; we can suppress it by enabling the doAODMerging flag.
So do that.

(Eventually should change this to the new job configuration...)