Adding draft translator for trans to pers names
To avoid warnings and reduce some code complexity, we want a translator
from transient container name (e.g. with _v3 in, or with DataVectr<>) to
persistent name (no _v3, or Container). This adds a little helper
function to do the translation, and uses it in RecExCommon -
successfully avoiding the warnings that are printed otherwise.