Fixing warnings in derivation running
First change to MCTruthClassifier setup fixes a warning that the same
tool is added to a tool handle array multiple times.

Second change fixes a warning that the input file metadata doesn't have
the currents available, and generates a warning that the geometry and
metadata differ in the current they see (which is concerning; I've
contacted the sim team about this).

Third change fixes a multiple add warning for the truth collections.

Fourth change guards against adding the same extrapolator multiple times
to the ToolSvc.