Adding metadata for event overlay
After a brief discussion on PAT-help, we decided it'd be helpful to
include metadata for data overlay in the xAOD metadata object. This
adds that metadata, as well as adding accessors and setters for booleans
that have been stored as chars.

In order to get the unit tests to behave, I had to make explicit the
types in the setters and getters. I suspect this is just because we have
too many options now, and the compiler is no longer able to
automatically tell what we want.