Fixing derivation crash: metadata come as a string
This fixes my misunderstanding with the metadata, that they can be
interpreted in a variety of ways automatically. Instead, this value
needs to be interpreted as a string. Since it's set in python, it's
guaranteed to have the value True or False (with that capitalization),
so this update should solve the problem related to access.