First pass cleaning some of MadGraphUtils
This picks away at some of the old stuff in MadGraphUtils that should be
cleaned up. There was quite a lot of if-else to do with dealing with old
MadGraph versions because of the flexibility of the setup, but in
release 21.6 and newer we really should never be using some of these
older MG versions, so the catches can be dropped. This is just a first
step, but does kill off a few lines of the thing.