Bugfix and test update for MGC
With the removal of a function, I missed that it was being used in the
helpers (this was a little evil: utils calling helpers calling utils).
Fixed now.

Adding to the tests some attempts to use the log parser for tests that
create an EVNT file. This should let us catch more errors simply than
those caught by panda.