Make configuration of the BLM sensitive detector consistent with BLM geometry building
In the BLM geometry configuration:
`DetectorDescription/GeoModel/AtlasGeoModel/python/InDetGM.py`
The BLM geometry is only built if the Pixel and BCM DetFlags are true.
Previously in the sensitive detector configuration the BLM SD was added
if the Pixel DetFlag, regardless of whether the BCM DetFlag was true.

This change commit makes the sensitive detector configuration consistent.
It should fix ATLASSIM-4427.