PHYS/LITE: Updating trigger matching to use 0.2 for taus
At the recommendation of the tau group, we should use 0.2 as the radius
for matching taus in the trigger. This MR updates the tau and tau+MET
triggers to use 0.2 for the trigger matching radius. The other triggers
continue with a radius of 0.1 (the default).

This relies on !29874 being accepted.