More updates for MGC tests
This resolves a case in simple_lhe_plotter where there was a comment in
the file that the script couldn't parse.

It also adds DCube tests to the first three MGC nightly jobs. If these
all work properly, then they'll be added to more in the future. If all
goes well, this will result in automatic webpages of plots comparing the
reference (from this weekend) to the current nightly.