WorkDir: Set VERSION to same value as parent project
This is a follow-up to MR !30019 where the `WorkDir` version was
hard-coded to `1.0.0`. But some of our cmake code relies on the
`CMAKE_PROJECT_VERSION` variable to be set to the same value as the
parent project. Achieve this by calling `project(...)` twice. First
without version, then a second time once we know the parent version
number.

Also remove the obsolete Doxygen work-around.