Updating and patching TRUTH derivation tests
Patch to FileMetaDataCreatorTool is for a problem that was introduced a
few MRs ago to do with the variety of accessors we now provide. This
adds the string literal operator to ensure that we get the right setter
and getter.

The test updates are two a new EVNT file, which corresponds to the mc16
default ttbar sample (replacing the old mc15 Powheg ttbar sample). This
will make use of some functionality that requires newer evgen files.

The update to the merging skeleton is to fix an issue that was
introduced because of file argument juggling. It simply reverses the
order of two tests, because one interfered with the other (now the test
that doesn't change anything comes first).