Adding collision mode for ComTimeRec
We sometimes want to run cosmic reco on collision events for the purpose
of long-lived particle analysis. In those cases, the timing of the
detector is the same as it is in collision mode (for the trigger, at
least), even though cosmic reconstruction is configured. This allows an
additional mode for the ComTimeRec algorithm, wherein it doesn't try to
find the cosmic track record in simulation, but assumes that we are
running something collision-like.