AthenaCommon: Some fixes for athena.py -s
The exclusion pattern for AthenaCommon was failing if AthenaCommon
was built in a local WorkArea.

sys.maxint doesn't exist in py3 --- change to maxsize.

With py3, we get _trace_include called with some internal bootstrap methods.
We were taking a huge amount of time trying to resolve these with FindFile.
Just ignore them instead.