WIP: Adding keep children strategy
The idea of this strategy is as a replacement for !27797

After chatting with John Chapman, we realized that we could set up a
truth strategy that only modifies the keep children setting of the truth
incident. This does *not* modify the decision about the strategy (keep
or throw away the vertex); it only modifies the setting of keep
children. This can be used on top of existing truth strategies to change
their behavior, for example by keeping all the decay products of an LLP
or keeping all the decay products of a hadronic interaction, even in a
mode where they would normally be dropped. Two example configurations
are given.