Providing help with annoying weight names
There are some generators that produce names that have spaces in them.
That can be quite frustrating for analysers that are even clever enough
to print out the weightnames, since they don't quite match what's in the
derivation. This puts out a warning in case the match is off by some
spaces, but still allows the event to continue (so that users can fix
their code).