Switching test to use a bare pointer
Trying to use a std::unique_ptr for the IAppMgr causes a crash in
finalization. Looks like the pointer is being cached somewhere in the
guts of Gaudi, so we get something like a double delete from the
SmartIF. Just moving it to a bare pointer (like the other tests in
athena seem to do) does the job.