genCLIDDB: Support running without input file and cleanup
Main change is to support running without a clid.db input file. If the
executable is run without `--input/-i` option, it now assumes that no
input file is used. To restore the previous behavior, one needs to
explicitly specify `-i clidb.db`.

In addition, major code cleanup and reduce the number of messages
printed when creating the CLID database.