Remove unused members from hole search
The logic that uses these is now moved to the boundary checking tool, so
there is no use in having them in the hole search anymore, they're
redundant. Best to get rid of them.

Note: This is a breaking commit. It is intended to be followed up by a
commit that fixes any configuration issues caused.