Adding Keep children strategy and changing defaults
Adding the KeepChildrenTruthStrategy from !30161 to master. There were a
few interface changes to the truth strategy base class that had to be
added. Adding a few extra helper functions in the python setup (on top
of what was made available in !30161).

Finally, tweaking the defaults for the LLP truth setups to use these
strategies, since I think that's what we really want for MC15 and later.