Update configuration of track summary tool used for muon triggers
In the muon trigger, we use the iPatFitter in sequences that build combind MS+ID muons and also sequences that only build MS muons (and don't run the ID tracking at all). The track summary tool used by the iPatFitter then needs to be configured correctly depending on whether we will have ID data or not available in the sequence. Since the combined muon track builder has the same issue and is all ready being confiugred to use an appropriate summary tool, we can just configure special instances of the iPatFitter to use the same version of the track summary tool.
Fixes ATR-20521.