Force McEventCollection merging to run before ID and Muon Digi in MT Overlay
The hard-scatter and background `McEventCollection` objects are merged in Overlay jobs.
This merging has to occur before the collection is accessed via the `HepMcParticleLink`
objects owned by ID and Muon SimHits. As the dependency is not explicit it is not picked
up by the MT scheduler. Therefore the dependency has to be added manually to affected Algorithms.
This is the same fix as !30525, but for the old-style rather than CA-based configuration.