Updates to run 2 muon trigger configuration
Updating the decoding tools used in Run 2 trigger to output to default collections (since these are used in quite a lot of places later on, and the collection names don't need to be different from the default).
Also no longer ignoring some misalgined CSC chambers as we did for Run 2 to make it easier to compare to the Run 3 triggers where we won't ignore these chambers by default.