Dropped the dependency on IncidentSvc from AthAnalysisAlgorithm
The dependency was redundant. AthAnalysisAlgorithm was registering itself as a listener
on three incidents, but in fact was listening only to BeginInputFile. The handler
of the latter incident was used only to set a redundant m_currentFile flag.

Also dropped the redundant currentFile method, which did not have any clients.