Dropped the dependency on TagInfo from LArG4GenShowerLib
This commit is a part of the effort to reduce the number of clients of TagInfo
in Athena to the bare minimum. After that the TagInfo handling mechanism should
be redesigned, in order to make it MT-compliant.

LArG4GenShowerLib used TagInfo only for querying it for the releaese information.
The latter information can be obtained from the environment variables instead.