Validation and updates to SUSY setup
This adds the first tests for the SUSY setup, for a gridpack production
and LHE file making off of the gridpack. It was an opportunity for
further cleanup and for the baby steps towards really pulling out the
'SUSY' part of MadGraphControl and isolating it. Adding a test for the
setup that I'm currently debugging.

At the same time, adding a -nojpeg flag to the output by default. This
should speed up production a bit, since we get to avoid the production
of Feynman diagram pictures that we don't need.