Fix for muon PRD collection names in Run 2 triggers
Trying to fix the PRD collection names in the Run 2 muon trigger configuration. We can't currently use the same collection names in the trigger and offline when running the offline and trigger in the same jobs (we run into the problems reported in ATR-20971).
However, there are many tools that depend on these collection names, and it's not that straight forward to change them all (and for code that will be obsolete soon, isn't really worth the effort). Since this is only a problem for the Run 2 trigger code, and this is only used for validation comparisons between Run 2 and Run 3 triggers where we don't run the trigger and offline reco in the same job, will do the following:
- Use the default offline names in jobs where we are not going to be running the offline and trigger at the same time, so we get correct Run2 trgger results
- Will use different names when we are running the trigger and offline at the same time and accept that the HLT muon trigger output is not 100% correct (it's a small effect anyway)