Fix for ATLASRECTS-5343
The CaloMuonLikelihoodTool should use ParticleCaloExtensionTool to extrapolate to the calo instead of the TrackEnergyInCaloTool.
This gives consistency with other tools (isolation, the MuonCreatorTool, etc.).
Additionally the ParticleCaloExtensionTool is written in a muon-friendly manner, making it a superior option.
The interface is modified to take a pointer to a TrackParticle, to take advantage of the ParticleCaloExtensionTool's capabilities.