The Great Renaming - renaming all the ID Trigger collections
The ID Trigger Track collections have been renamed using the new consistent
naming convention as discussed in the TGM of 2019-08-28 (https://indico.cern.ch/event/843886/)

https://indico.cern.ch/event/843886/contributions/3545029/attachments/1898633/3133218/sutt-idcollections-2019-08-28.pdf

This should include all the changes to the configuration of the ID Trigger
and the menu code for the signatures, including the signature monitoring.

The vertex collections still need to be renamed - this will be done in the
future once the track renaming is complete.