disable QuickAna in analysis releases
Since QuickAna doesn't seem maintained any more, the best solution may
be to take it out of the release, but keep it in the repository (for
now).  That way if somebody steps up to maintain it we can reinclude
it if needed, or if nobody misses it this would be the first step in
removing it completely.