CaloRec: Remove StatusCode return from declareConstant.
ToolWithConstantsMixin::declareConstant was returning a StatusCode.
However, this is used in constructors, so in practice, the return
value was never being checked.  Pending changes to Gaudi will start
giving warnings about this.

There's no point in returning a StatusCode if it can't reasonably be checked,
so just change declareConstant (and add_constant) to be void.