Two new files added. RootHelpers.cxx and RootHelpersh . This files contain the interpolate functions that I added in Helpers.cxx and Helpers.h . The files Helpers.cxx and Helpers.h have been set as in the original version (21.2).
The UncertaintyHistogram.cxx and UncertaintyHistogram.h files have been modify to use the Interpolate and FindBin function defined in RootHelpers.cxx

The structure to read the histograms and the .config files from the
CalibArea-8 as in JetUncertainties has been added (but the final
function is commeted). Final tests have to be performed once they can be
found in the CalibArea.

After those final test with the files in the CalibArea the Tool will be
ready for the merge request