ActsGeometry: clang 9 fixes
There's an issue with clang 9 / boost that gives link errors

```
  multiple definition of `boost::hana::curry_detail::curry_or_call<0ul>'
```

if some acts headers are included more than once in a library.

We get this from ActsExtrapolationTool.  Work around this by moving all
references to the affected acts headers from the header to the cxx file.

Also fix a number of clang warnings.