Replace track parameter related memory management in the extrapolator.
Now the track parameters are managed with smart pointers and stored in the
current extrapolation cache. The smart pointers support externally managed,
released and shared track parameters. All non external and not-released track
parameters are deleted together with the current extrapolation cache. The smart
pointers ensure that all track parameters are deleted which are not returned
to the caller and which are not managed externally, and that the deletion is
delayed until all references have vanished.

This addresses ATLASRECTS-5340.