Pythia8_i: fix enhance hook to manually take into account event weight
We have to manually multiply the enhancement weight (`enhanceHooks->getEnhancedEventWeight()`) to the event weight. Unfortunately in addition, this weight is not properly initialised in Pythia, so we can't just generally do this for all UserHooks but have to specifically do it only for enhancing hooks.

Fixes AGENE-1802 and amends af1f917b