Fixing three Pythia8/MadGraph issues in 21.6
First issue: we need to make sure there is a good default for the PDG
IDs to merge in SUSY jobs (the default is now set in the preinclude).

Second issue: we should make sure that if the user is loading a model
that isn't the standard SUSY model, but still wants to use the SUSY
infrastructure, the job doesn't crash because of our attempts to provide
helpful definitions.

Third issue: Historically, Pythia8 had a problem that if "guess" was
used for merging, it would crash during initialization. That was worked
around by setting the merging to a dummy process during initialization
and then setting the code back to using guess. Recently (Pythia8 244 at
least), merging stopped passing any events. It appears that by removing
the work-around, "guess" does not crash the job during initialization
(so the Pythia8 people fixed that bug), and the merging now works (so
there must have been some change to initialization that requires guess
to be set already there).