Removing obsolete files
Looking through these, there are more modern ways they should be done in
job options. I haven't identified any use of these functions in anything
running in release 21.6, so removing them for now. If we need to add
them back in later, we always can do, but I expect we won't need to.