Add full implementation of new RegSelTool Conditions data access for Pixel and SCT
This includes all the fully functional changes for the new conditions data access
RegSelTool access to replace the egregious old RegSelSvc class

New classes are included:
```
   * RegSelCondData    a generic wrapper class for conditions data
   * RegSelLUTCondData the specific instance of the above wrapper for the RegSelSiLUT
                        conditions data for the ID And MS lookup tables
   * SiRegSelCondAlg    the new conditions algorithms to add the conditions data
```
Modifications are also included for the FastTrackFinder and the
TrigSpacePointConversionTool to replace the use of the RegSelSvc for the Pixel and
SCT in the SpacepointConversion tool by way of example - it is not a full replacement
in this package.

New configuration functions are provided in
```
	RegSelToolConfig.py
```
such that components can include the relevant RS tools using fragments such as
```
  from RegionSelector.RegSelToolConfig import makeRegSelTool
  spTool.RegSel_Pixel = makeRegSelTool("Pixel")
  spTool.RegSel_SCT   = makeRegSelTool("SCT")
```
The setup functions register the corresponding conditions algorithms when called
ensuring that they are registered once, and once only.

The code to use the RegSelSvc in this tool has been retained, but has been commented
out, since this change is still very much ongoing. The cleanup to remove this older
code will come along with the full replacement of the RegSelSvc in the full package
(and others) once the remaining RegSelTool developments are more complete.

Changes for the TRT and Muon Spectrometer region selector code will come once these
initial changes have been merged.