Add new RegSelTool for the TRT also
This adds the TRT_RegSelCondAlg class which is a cut and paste of the
TRT_RegionSelectorTable code, but wrapped in a Conditions algorithm.

Also the
```
makeRegSelTool_TRT
```
configuration function has been added to the
```
RegSelToolConfig.py
```
configuration file. The other configuration functions will be added to
this when they have been implemented.

These changes for the TRT complete the implementation of the inner detector
RegSelTools.