Create InDet calo ROI seeds if they are used for TRT back-tracking.
The calo seeds have only been created if brem recovery was on. Thus, in
cases the brem recovery was off (e.g. solenoid off), and calo seeds were
still needed for TRT back tracking, the job would fail. Now, the calo
seeds are created if either brem recovery is on, or TRT back tracking
is on and requires calo seeds.