Properly configure track summary tools in muon trigger
Updating the configuration of the track summary tools used in the muon trigger for the muon creation and ctrack cleaning so that we use the MuonTrackSummaryTool when building MS-only muons and have no ID tracks, and the combined muon track summary tool when we do have ID tracks. This fixes the warnings reported in ATR-20521.