Initialise muon conditions data handles only if they are needed
Changing some of the conditions data handles to be initialised only if they are actually going to be used (ie initialize conditions needed only for data only if we're running on data and similar for conditions that are needed only for MC).
In cases where the data does not exist in the first place, this will help avoid warnings along the lines of WARNING Will not auto-load proxy for non-EventStore object:  ( 'CondAttrListCollection' , 'ConditionStore+/CSC/DCS/LAYERSTATE' ).