Tweaks to MGC to deal with cases
The run card in MG is not case-sensitive, and if a variable is set
multiple times it takes the last setting. That means up to now if
```iseed``` and ```iSeed``` were set, the one set second was used. That
saved some users who used different capitalization from what's found in
the run card, but is confusing. Now we rely on setting everything to
lower case and doing a case-insensitive match, and then using the
capitalization specified by MG5 if a setting exists. This should help
clarity for users.