MGC SUSY bug fixes for madspin and jo
Fixing two issues in MadGraphControl. The first is that the SUSY
pre/post include setup for systematics was relying on the job options
name, which needs to be grabbed a different way. The second is that the
SUSY setup wasn't using the MadSpin card it was given. Both should be
fixed here.