AthDerivation Ninja fixes.
Taught xAODRootAccess to only install its checkxAOD.py script when
building a standalone analysis release, in order not to clash with
the script of the same name coming from PyUtils.

Removed the test_mc16PHYSVAL.sh ART test script from
DerivationFrameworkPhysicsValidation, as it should rather be taken
from DerivationFrameworkART now.

Fixed a typo in the CMake configuration of TauHistUtils.