Comparison between JetUncertaintiesTool Calo weights and the JetCalibrationTool Calo weights. Now we test that neither the Calo nor the TA mass have a value lower than 1 GeV.
If one of them have that low mass, it means that the weights do not
match and we use the JetCalibration weights to rejoin the jets into the
COmbined mass. If the mass is higher, it means that both weights should
match and then we can compute the new weights using JetUncertainties
(the weights are "new" and different from the original ones because the
mass has been smeared). If the systematic do not affect our jet, we
always use the JetCalib weights to avoid any (small) computing
non-clusure