Make the ZFinder::findZ() method const
Changes to allow the findZ() method to work without writing to any class variables
Several accessors were removed which depended on class variables being updated,
such as the method to access the actual z histograms after the processing.

Since the processing is now all done with local variables, accessing the internal
state won't be possible outside the function itself unless we store the internal
state spacifically.

Fortunately, the results appear identical with the original version so we should
be able to live without this for the time being.

The actual code itself hasn't been changed much - some of the formating has been
slightly improved - but the code itself has been left as it is, in order to
maintain the same fuctionality, so any changes to the actual functionality of
the code would need further studies and a separate merge request.