DataModelRoot: More adjustment of thread-safe markings.
The TScopeAdapter constructor was marked as not-thread-safe, due to its
manipulation of the default error level.  However, calls to this constructor
were not being properly checked due to an issue with the checker.

That issue is now being fixed, but that means that we now have some warnings
from these calls.

Clean up a bit by splitting the constructor in two.  One doesn't have the
quiet option and can be considered thread-safe, while one which takes
a quiet argument is still marked as not thread-safe.  Adjust calls
appropriately.

We can also clean up the implementation a bit.