AthenaPoolUtilities: Avoid copying AttributeList.
Copying a coral::AttributeList is not thread-safe.
Change AthenaAttributeList::coralList() to return a reference
rather than making a copy.

Add ATLAS_NOT_THREAD_SAFE declarations for some methods that do need
to make copies of AttributeLists.

Change CondAttrListCollection::add to return a value, to allow the possibility
of suppressing warnings about calls to it.