Update digit container names used for L1 muon simulation
Updating the names of the RPC and TGC digit containers created for the L1 muon simulation to be unique in order to avoid trying to write to the all ready existing containers. Also removing a couple of instances where we include the MuonRDOToDigit to digit job options, since this appeared to be scheduling the RDO to DIGIT conversion twice and it is only needed to run as part of the L1 simulation.