Updates to the RegSelTool implementation including addition of MDT tables
These include several updates to the RegSelTool implementation with respect
to merge request !30981

Firstly better memory handling for the RegSelLUTCondData where the
payload is now stored using a unique_ptr to better avoid memory leaks.

Additionally there is an new class for the MDT_RegSelCondAlg which
uses the new MDT cabling conditions data to create the table for the
RegSelTool.

The RegSelToolConfig.py now has an improved private builder function,
whoch is used to provide public builders for the Pixel, SCT, TRT and MDT.

The SCT Cabling conditions data are now used for the EventIDRange for
the SCT table, rather than using the pixel cabling.

Some of the old RegSelSvc code has been retained, but commented since
we are still developing.

One additional factor regarding the MDT table creation code: The
createTable() method is essentially a cut and past of the original
MDT code from the MDT_RegionSelectorTable class. I did not originally
write this code, and don't want to touch it as that would require a
great deal of validation beyond the scope of the changes for this
merge request. There are a good many places where that code itself
could be improved, but that should be left for the Muon developers
for a later day.

For this code we are implementing the framework to provide the full
functionality for the new RegSelTool implementation so there will be
quite a lot of copying of code from the older implementation over
the coming weeks for the additional Muon and Calorimeter tables.