Accommodate multiple collections with ReadHandle
When working with two collections -- one for hard-scatter and another
for pile-up -- we need to look at different `StoreGate`s. Previously,
when the `AthAlgTool::EvtStore` property was overridden in JOs, this
would also effect the global SG key.

This patch ensures that the `TruthEvent` is retrieved from the default
`StoreGate` instance -- i.e. that named `StoreGateSvc` -- rather than
the `OriginalEvent_SG` when using multiple collections.