tauRecTools: terminate tau reconstruction for invaild seed link
Previously, the treatment for invalid seed link is different for
different tools. Some tools return StatusCode::SUCCESS and print out a
WARNING; some return StatusCode::FAILURE and print out an ERROR. The
point is that the seed link should be always valid. In case it is
invalid, there must be some problems. Then, tau reconstruction should
be terminated. It is fixed now.