CaloConditions: Conditions objects for lumi-dependent pileup offset correction.
The cell-by-cell luminosity-dependent offset correction for adjacent bunches
was implemented as an algorithm (CaloBCIDAvgAlg), but this took too long
to be usable in the trigger.  We can, however, speed it up dramatically
by doing much of the work in advance and vectorizing the remainder.
The processed data depends on the cell pulse shapes and OFCs as well
as the per-bunch luminosity.  The first two of these are saved in
CaloBCIDCoeffs and the third (because it has a different lifetime)
in CaloBCIDLumi.